+++ CHANGELOG.md
@@ -1,6 +1,7 @@
-# Catalyst SD-WAN Lab 2.0.15 [unreleased]
+# Catalyst SD-WAN Lab 2.0.16 [unreleased]
 
 - In setup task, add option to delete already defined software images
+- Fix issue where add task might fail for SD-WAN Manager 20.18 and higher
 
 # Catalyst SD-WAN Lab 2.0.15 [Feb 25, 2025]