+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# Catalyst SD-WAN Lab 2.0.16 [unreleased]
+
+- In setup task, add option to delete already defined software images
+- Fix issue where add task might fail for SD-WAN Manager 20.18 and higher
+- In setup task, change behaviour when --list is specified, print software version and exit
+- Update validator cloud-init to support 20.18 release
+
 # Catalyst SD-WAN Lab 2.0.15 [Feb 25, 2025]
 
 - Fixed a problem where the backup task might fail with a "KeyError: 'label'"