+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - In setup task, add option to delete already defined software images
 - Fix issue where add task might fail for SD-WAN Manager 20.18 and higher
 - In setup task, change behaviour when --list is specified, print software version and exit
+- In setup task, verify license and print error that the tool is not supported on CML-Free as it doesn't support required number of nodes
 - Update validator cloud-init to support 20.18 release
 - In restore task, add option to change the software version of control components and SD-WAN/SD-Routing edges during restore
 - Increase SD-WAN Manager CPU count from 8 to 10