+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Fix issue where add task fails with traceback if user removed default device templates/configuration groups
 - Change Gateway node type from c8000v to IOL XE
 - Change transports node type from unmanaged_switch to IOL L2 XE
+- Add support for IPv6 and dual stack overlays
 
 # Catalyst SD-WAN Lab 2.0.15 [Feb 25, 2025]