+++ CHANGELOG.md
@@ -1,5 +1,10 @@
+# Catalyst SD-WAN Lab 2.1.1 [Jun 18, 2025]
+
+- Print error when user tries to use the SD-WAN Lab Tool 2.1.x or higher against CML 2.6.x
+
 # Catalyst SD-WAN Lab 2.1.0 [Jun 17, 2025]
 
+- Drop support for CML 2.6.x
 - Add support for IPv6 and dual stack overlays
 - Change Gateway node type from c8000v to IOL XE
 - Change transports node type from unmanaged_switch to IOL L2 XE